+++ src/main/java/io/gravitee/policy/apikey/ApiKeyPolicy.java
@@ -15,39 +15,55 @@
  */
 package io.gravitee.policy.apikey;
 
+import static io.gravitee.gateway.reactive.api.policy.SecurityToken.TokenType.API_KEY;
+import static io.gravitee.gateway.reactive.api.policy.SecurityToken.TokenType.MD5_API_KEY;
+
 import io.gravitee.common.http.GraviteeHttpHeader;
 import io.gravitee.common.http.HttpStatusCode;
 import io.gravitee.gateway.api.service.ApiKey;
 import io.gravitee.gateway.api.service.ApiKeyService;
 import io.gravitee.gateway.reactive.api.ExecutionFailure;
 import io.gravitee.gateway.reactive.api.context.ContextAttributes;
-import io.gravitee.gateway.reactive.api.context.HttpExecutionContext;
-import io.gravitee.gateway.reactive.api.context.HttpRequest;
-import io.gravitee.gateway.reactive.api.policy.SecurityPolicy;
+import io.gravitee.gateway.reactive.api.context.base.BaseExecutionContext;
+import io.gravitee.gateway.reactive.api.context.http.HttpPlainExecutionContext;
+import io.gravitee.gateway.reactive.api.context.http.HttpPlainRequest;
+import io.gravitee.gateway.reactive.api.context.kafka.KafkaConnectionContext;
 import io.gravitee.gateway.reactive.api.policy.SecurityToken;
+import io.gravitee.gateway.reactive.api.policy.http.HttpSecurityPolicy;
+import io.gravitee.gateway.reactive.api.policy.kafka.KafkaSecurityPolicy;
 import io.gravitee.policy.apikey.configuration.ApiKeyPolicyConfiguration;
 import io.gravitee.policy.v3.apikey.ApiKeyPolicyV3;
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Maybe;
+import java.security.SecureRandom;
 import java.util.Date;
 import java.util.Optional;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import java.util.function.Function;
+import javax.security.auth.callback.Callback;
+import javax.security.auth.callback.NameCallback;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.kafka.common.security.plain.PlainAuthenticateCallback;
+import org.apache.kafka.common.security.scram.ScramCredential;
+import org.apache.kafka.common.security.scram.ScramCredentialCallback;
+import org.apache.kafka.common.security.scram.internals.ScramFormatter;
+import org.apache.kafka.common.security.scram.internals.ScramMechanism;
 
 /**
  * @author David BRASSELY (david.brassely at graviteesource.com)
  * @author Jeoffrey HAEYAERT (jeoffrey.haeyaert at graviteesource.com)
  * @author GraviteeSource Team
  */
-public class ApiKeyPolicy extends ApiKeyPolicyV3 implements SecurityPolicy {
+@Slf4j
+public class ApiKeyPolicy extends ApiKeyPolicyV3 implements HttpSecurityPolicy, KafkaSecurityPolicy {
 
     static final String ATTR_API_KEY = ContextAttributes.ATTR_PREFIX + "api-key";
     static final String ATTR_INTERNAL_API_KEY = "api-key";
+    static final String ATTR_INTERNAL_MD5_API_KEY = "md5-api-key";
     static final String API_KEY_HEADER_PROPERTY = "policy.api-key.header";
     static final String API_KEY_QUERY_PARAMETER_PROPERTY = "policy.api-key.param";
     static final String DEFAULT_API_KEY_QUERY_PARAMETER = "api-key";
     static final String DEFAULT_API_KEY_HEADER_PARAMETER = GraviteeHttpHeader.X_GRAVITEE_API_KEY;
-    private static final Logger log = LoggerFactory.getLogger(ApiKeyPolicy.class);
     static String API_KEY_HEADER, API_KEY_QUERY_PARAMETER;
     private final boolean propagateApiKey;
 
@@ -62,12 +78,12 @@ public class ApiKeyPolicy extends ApiKeyPolicyV3 implements SecurityPolicy {
     }
 
     @Override
-    public Maybe<SecurityToken> extractSecurityToken(HttpExecutionContext ctx) {
+    public Maybe<SecurityToken> extractSecurityToken(HttpPlainExecutionContext ctx) {
         final Optional<String> apiKeyOpt = extractApiKey(ctx);
         if (apiKeyOpt.isPresent()) {
             String apiKey = apiKeyOpt.get();
             if (apiKey.isBlank()) {
-                return Maybe.just(SecurityToken.invalid(SecurityToken.TokenType.API_KEY));
+                return Maybe.just(SecurityToken.invalid(API_KEY));
             }
             ctx.setInternalAttribute(ATTR_INTERNAL_API_KEY, apiKey);
             return Maybe.just(SecurityToken.forApiKey(apiKey));
@@ -97,11 +113,11 @@ public class ApiKeyPolicy extends ApiKeyPolicyV3 implements SecurityPolicy {
     }
 
     @Override
-    public Completable onRequest(final HttpExecutionContext ctx) {
+    public Completable onRequest(final HttpPlainExecutionContext ctx) {
         return handleSecurity(ctx);
     }
 
-    private Completable handleSecurity(final HttpExecutionContext ctx) {
+    private Completable handleSecurity(final HttpPlainExecutionContext ctx) {
         return Completable
             .defer(() -> {
                 try {
@@ -116,18 +132,8 @@ public class ApiKeyPolicy extends ApiKeyPolicyV3 implements SecurityPolicy {
                         .getComponent(ApiKeyService.class)
                         .getByApiAndKey(ctx.getAttribute(ContextAttributes.ATTR_API), requestApiKey.get());
 
-                    if (apiKeyOpt.isPresent()) {
-                        ApiKey apiKey = apiKeyOpt.get();
-
-                        // Add data about api-key, plan, application and subscription into the execution context.
-                        ctx.setAttribute(ContextAttributes.ATTR_APPLICATION, apiKey.getApplication());
-                        ctx.setAttribute(ContextAttributes.ATTR_SUBSCRIPTION_ID, apiKey.getSubscription());
-                        ctx.setAttribute(ContextAttributes.ATTR_PLAN, apiKey.getPlan());
-                        ctx.setAttribute(ATTR_API_KEY, apiKey.getKey());
-
-                        if (isApiKeyValid(ctx, apiKey)) {
-                            return Completable.complete();
-                        }
+                    if (this.handleApiKey(apiKeyOpt, ctx, (apiKey -> true))) {
+                        return Completable.complete();
                     }
                 } catch (Throwable t) {
                     log.warn("An exception occurred when trying to verify apikey.", t);
@@ -138,22 +144,22 @@ public class ApiKeyPolicy extends ApiKeyPolicyV3 implements SecurityPolicy {
             .doOnTerminate(() -> cleanupApiKey(ctx));
     }
 
-    private boolean isApiKeyValid(HttpExecutionContext ctx, ApiKey apiKey) {
-        return !apiKey.isRevoked() && (apiKey.getExpireAt() == null || apiKey.getExpireAt().after(new Date(ctx.request().timestamp())));
+    private boolean isApiKeyValid(BaseExecutionContext ctx, ApiKey apiKey) {
+        return !apiKey.isRevoked() && (apiKey.getExpireAt() == null || apiKey.getExpireAt().after(new Date(ctx.timestamp())));
     }
 
-    private Completable interrupt401(HttpExecutionContext ctx, String key) {
+    private Completable interrupt401(HttpPlainExecutionContext ctx, String key) {
         return ctx.interruptWith(new ExecutionFailure(HttpStatusCode.UNAUTHORIZED_401).key(key).message(API_KEY_UNAUTHORIZED_MESSAGE));
     }
 
-    private Optional<String> extractApiKey(HttpExecutionContext ctx) {
+    private Optional<String> extractApiKey(HttpPlainExecutionContext ctx) {
         // 1_ First, check if already resolved.
         String apiKey = ctx.getInternalAttribute(ATTR_INTERNAL_API_KEY);
         if (apiKey != null) {
             return Optional.of(apiKey);
         }
 
-        final HttpRequest request = ctx.request();
+        final HttpPlainRequest request = ctx.request();
 
         // 2_ Second, search in HTTP headers
         if (request.headers().contains(API_KEY_HEADER)) {
@@ -176,11 +182,109 @@ public class ApiKeyPolicy extends ApiKeyPolicyV3 implements SecurityPolicy {
         return Optional.ofNullable(apiKey);
     }
 
-    private void cleanupApiKey(HttpExecutionContext ctx) {
+    private void cleanupApiKey(HttpPlainExecutionContext ctx) {
         if (!propagateApiKey) {
             ctx.request().headers().remove(API_KEY_HEADER);
             ctx.request().parameters().remove(API_KEY_QUERY_PARAMETER);
         }
         ctx.removeInternalAttribute(ATTR_INTERNAL_API_KEY);
+    }
+
+    @Override
+    public Maybe<SecurityToken> extractSecurityToken(KafkaConnectionContext ctx) {
+        Callback[] callbacks = ctx.callbacks();
+        for (Callback callback : callbacks) {
+            if (callback instanceof NameCallback nameCallback) {
+                // With SASL_PLAIN or SCRAM, we expect the username to be a md5 hash of the api-key, for security and privacy.
+                String md5ApiKey = nameCallback.getName();
+                if (md5ApiKey == null || md5ApiKey.isBlank()) {
+                    md5ApiKey = nameCallback.getDefaultName();
+                }
+                if (md5ApiKey == null || md5ApiKey.isBlank()) {
+                    return Maybe.just(SecurityToken.invalid(MD5_API_KEY));
+                }
+
+                ctx.setInternalAttribute(ATTR_INTERNAL_MD5_API_KEY, md5ApiKey);
+                return Maybe.just(SecurityToken.forMD5ApiKey(md5ApiKey));
+            }
+        }
+        return Maybe.empty();
+    }
+
+    @Override
+    public Completable authenticate(KafkaConnectionContext ctx) {
+        return Completable
+            .defer(() -> {
+                String md5ApiKey = ctx.getInternalAttribute(ATTR_INTERNAL_MD5_API_KEY);
+                final Optional<ApiKey> apiKeyOpt = ctx
+                    .getComponent(ApiKeyService.class)
+                    .getByApiAndMd5Key(ctx.getAttribute(ContextAttributes.ATTR_API), md5ApiKey);
+
+                if (
+                    this.handleApiKey(
+                            apiKeyOpt,
+                            ctx,
+                            apiKey -> {
+                                Callback[] callbacks = ctx.callbacks();
+                                for (Callback callback : callbacks) {
+                                    if (callback instanceof PlainAuthenticateCallback plainAuthenticateCallback) {
+                                        plainAuthenticateCallback.authenticated(true);
+                                        return true;
+                                    } else if (callback instanceof ScramCredentialCallback scramCredentialCallback) {
+                                        ScramCredential scramCredential = createScramCredential(
+                                            apiKey.getKey(),
+                                            ScramMechanism.forMechanismName(ctx.saslMechanism())
+                                        );
+                                        scramCredentialCallback.scramCredential(scramCredential);
+                                        return true;
+                                    }
+                                }
+                                return false;
+                            }
+                        )
+                ) {
+                    return Completable.complete();
+                }
+                return Completable.error(new Exception(API_KEY_INVALID_KEY));
+            })
+            .doOnTerminate(() -> cleanupApiKey(ctx));
+    }
+
+    private void cleanupApiKey(KafkaConnectionContext ctx) {
+        ctx.removeInternalAttribute(ATTR_INTERNAL_MD5_API_KEY);
+    }
+
+    @SneakyThrows
+    private ScramCredential createScramCredential(String password, ScramMechanism mechanism) {
+        // Number of iterations for PBKDF2
+        int iterations = 4096;
+
+        // Generate a random salt (typically 16 bytes)
+        byte[] salt = new byte[25];
+        SecureRandom secureRandom = new SecureRandom();
+        secureRandom.nextBytes(salt);
+
+        // ScramFormatter helps derive the storedKey and serverKey
+        ScramFormatter formatter = new ScramFormatter(mechanism);
+
+        // Generate storedKey and serverKey using the formatter
+        return formatter.generateCredential(salt, formatter.saltedPassword(password, salt, iterations), iterations);
+    }
+
+    private boolean handleApiKey(Optional<ApiKey> apiKeyOpt, BaseExecutionContext ctx, Function<ApiKey, Boolean> handleIfApiKeyIsValid) {
+        if (apiKeyOpt.isPresent()) {
+            ApiKey apiKey = apiKeyOpt.get();
+
+            // Add data about api-key, plan, application and subscription into the execution context.
+            ctx.setAttribute(ContextAttributes.ATTR_APPLICATION, apiKey.getApplication());
+            ctx.setAttribute(ContextAttributes.ATTR_SUBSCRIPTION_ID, apiKey.getSubscription());
+            ctx.setAttribute(ContextAttributes.ATTR_PLAN, apiKey.getPlan());
+            ctx.setAttribute(ATTR_API_KEY, apiKey.getKey());
+
+            if (isApiKeyValid(ctx, apiKey)) {
+                return handleIfApiKeyIsValid.apply(apiKey);
+            }
+        }
+        return false;
     }
 }